feat(api): add create/update/delete endpoints for organizations

+++ libs/project/src/project.service.ts
@@ -1,6 +1,7 @@
 import { CreateProjectDto } from '@app/project/dto/create-project.dto';
 import { Project } from '@app/project/entities/project.entity';
 import { ProjectMember } from '@app/project/entities/project-member.entity';
+import { OrganizationRole } from '@app/project/enum/organization-role.enum';
 import { ProjectRole } from '@app/project/enum/project-role.enum';
 import { OrganizationService } from '@app/project/organization.service';
 import { UserDto } from '@app/user/dto/user.dto';
@@ -8,7 +9,11 @@ import { UserService } from '@app/user/user.service';
 import { EntityRepository } from '@mikro-orm/core';
 import { InjectRepository } from '@mikro-orm/nestjs';
 import { wrap } from '@mikro-orm/postgresql';
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  ForbiddenException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 
 @Injectable()
 export class ProjectService {
@@ -58,11 +63,18 @@ export class ProjectService {
     return project;
   }
 
-  findOne(id: string, user: UserDto): Promise<Project | null> {
+  findOne(
+    id: string,
+    user: UserDto,
+    requiredRoles: ProjectRole[] = [ProjectRole.MEMBER, ProjectRole.OWNER],
+  ): Promise<Project | null> {
     return this.projectRepository.findOne(
       {
         id,
         members: {
+          role: {
+            $in: requiredRoles,
+          },
           user: {
             id: user.id,
           },
@@ -75,7 +87,11 @@ export class ProjectService {
   }
 
   async remove(id: string, user: UserDto): Promise<void> {
-    const project = await this.findOne(id, user);
+    const project = await this.findOne(id, user, [ProjectRole.OWNER]);
+
+    if (!project) {
+      throw new ForbiddenException();
+    }
 
     await this.projectRepository.getEntityManager().removeAndFlush(project);
   }
@@ -85,29 +101,4 @@ export class ProjectService {
 
     return project;
   }
-
-  // async findByUserRole(userProjectRole: UserProjectRole): Promise<Project> {
-  //   return this.projectRepository.findOne(
-  //     { userRoles: userProjectRole },
-  //     { populate: ['userRoles'] },
-  //   );
-  // }
-  //
-  // async findById(id: string, user: User): Promise<Project | undefined> {
-  //   return this.projectRepository
-  //     .createQueryBuilder('p')
-  //     .innerJoin('p.userRoles', 'upr')
-  //     .where('p.id = :id', { id })
-  //     .andWhere('upr.userId = :userId', { userId: user.id })
-  //     .getOne();
-  // }
-  //
-  // async findAllByOrganization(organization: Organization): Promise<Project[]> {
-  //   return this.projectRepository
-  //     .createQueryBuilder('p')
-  //     .where('p.organizationId = :organizationId', {
-  //       organizationId: organization.id,
-  //     })
-  //     .getMany();
-  // }
 }
